+++ lib/app.js
@@ -20,9 +20,8 @@ const authStudent = require('./routes/authStudent');
 const student = require('./routes/Student');
 const teacher = require('./routes/Teacher');
 const course = require('./routes/Course');
-const search = require('./routes/search');
 const flashcardSet = require('./routes/flashcardSet');
-const returnedFlashcardSet = require('./routes/returnedFlashcardSet');
+const flashcards = require('./routes/flashcards');
 
 if (process.env.NODE_ENV !== 'production') {
     app.use(checkDb);
@@ -30,12 +29,11 @@ if (process.env.NODE_ENV !== 'production') {
 
 app.use('/api/authTeacher', authTeacher);
 app.use('/api/authStudent', authStudent);
-app.use('/api/Student', ensureAuth, student);
-app.use('/api/Teacher', ensureAuth, teacher);
+app.use('/api/student', ensureAuth, student);
+app.use('/api/teacher', ensureAuth, teacher);
 app.use('/api/courses', ensureAuth, course);
-app.use('/api/search', ensureAuth, search);
-app.use('/api/flashcardSet', ensureAuth, flashcardSet);
-app.use('/api/search/set', ensureAuth, returnedFlashcardSet);
+app.use('/api/flashcards', ensureAuth, flashcards);
+app.use('/api/flashcardSets', ensureAuth, flashcardSet);
 
 
 app.use((req, res) => {
