+++ lib/models/user.js
@@ -20,18 +20,22 @@ const schema = new Schema({
     ]
 });
 
-schema.static('exists', function(query) {
+schema.static('exists', function (query) {
     return this.find(query)
         .count()
         .then(count => (count > 0));
 });
 
-schema.method('generateHash', function(password) {
+schema.method('generateHash', function (password) {
     this.hash = bcrypt.hashSync(password, 8);
 });
 
-schema.method('comparePassword', function(password) {
+schema.method('comparePassword', function (password) {
     return bcrypt.compareSync(password, this.hash);
 });
 
-module.exports = mongoose.model('Student', schema);
+// you can literally share the schema...
+module.exports = {
+    Teacher: mongoose.model('Teacher', schema),
+    User: mongoose.model('User', schema)
+};
