+++ lib/routes/Course.js
@@ -3,6 +3,8 @@ const router = Router();
 const Course = require('../models/course');
 const FlashcardSet = require('../models/flashcardSet');
 
+// these all need to limit for user who is making request
+
 router
     .get('/', (req, res, next) => {
         Course.find(req.query)
