+++ lib/routes/authStudent.js
@@ -5,6 +5,7 @@ const ensureAuth = require('../auth/ensure-auth')();
 const tokenService = require('../auth/token-service');
 const bodyParser = require('body-parser').json();
 
+// Don't repeat files like this.
 function hasEmailAndPassword(req, res, next) {
     const user = req.body;
     if (!user || !user.email || !user.password) {
